refactor: 업로드 워커 및 관련 훅 관심사 분리
- useUploadWorker.ts를 4개 모듈로 분리
  - shared-worker-manager.ts: SharedWorker 싱글톤 및 리스너 관리
  - service-worker-manager.ts: Service Worker 등록 및 Background Sync
  - useWebWorker.ts: Web Worker 생성/메시지 핸들링/네트워크 동기화
  - useUploadWorker.ts: 조합(orchestration)만 담당
- upload.worker.ts를 3개 모듈로 분리
  - rate-limiter.ts: Token Bucket Rate Limiter, Throttled Stream
  - upload-task.ts: S3 업로드, performUpload 로직
  - upload.worker.ts: 큐 처리, 메시지 핸들링
- Worker cleanup 시 terminate 전에 스로틀 해제 메시지 전송하도록 개선
- useGameStream의 업로드 에러를 console.error에서 toast로 변경
