+++ src/features/game-streaming-session/workers/upload.worker.ts
@@ -1,169 +1,38 @@
-import { postInputLogs, postPresignedUrl, postUploadComplete } from '../api';
+/**
+ * Upload Worker
+ *
+ * 세그먼트 업로드 큐 처리 및 메시지 핸들링.
+ */
 import type { StreamHealthState } from '../hooks/stream/useStreamHealth';
 import { UploadQueue } from '../lib/upload/upload-queue';
-import type { InputLog } from '../types';
+import { UploadRateLimiter } from './rate-limiter';
+import { performUpload, type UploadTask } from './upload-task';
 import type {
   UploadWorkerCommand,
   UploadWorkerEvent,
   UploadWorkerSegmentPayload,
 } from './upload-worker.types';
 
-interface UploadTask {
-  sessionId: string;
-  sequence: number;
-  segment: UploadWorkerSegmentPayload['segment'];
-  contentType: string;
-  blob: Blob;
-  logs: InputLog[];
-  remoteSegmentId?: string;
-  s3Url?: string;
-  s3Uploaded: boolean;
-  completeNotified: boolean;
-  logsUploaded: boolean;
-}
-
-class UploadRateLimiter {
-  private rateBytesPerSecond: number | null = null;
-  private capacity = 0;
-  private tokens = 0;
-  private lastRefill = Date.now();
-
-  setRate(rateBps: number | null): void {
-    if (!rateBps || rateBps <= 0) {
-      this.rateBytesPerSecond = null;
-      this.capacity = 0;
-      this.tokens = 0;
-      return;
-    }
-
-    this.rateBytesPerSecond = rateBps / 8;
-    this.capacity = this.rateBytesPerSecond;
-    this.tokens = this.capacity;
-    this.lastRefill = Date.now();
-  }
-
-  isEnabled(): boolean {
-    return this.rateBytesPerSecond != null;
-  }
-
-  private refill(): void {
-    if (!this.rateBytesPerSecond) return;
-    const now = Date.now();
-    const elapsedSeconds = (now - this.lastRefill) / 1000;
-    if (elapsedSeconds <= 0) return;
-    this.lastRefill = now;
-    this.tokens = Math.min(
-      this.capacity,
-      this.tokens + this.rateBytesPerSecond * elapsedSeconds
-    );
-  }
-
-  async consume(bytes: number, signal?: AbortSignal): Promise<void> {
-    if (!this.rateBytesPerSecond) return;
-    if (bytes <= 0) return;
-
-    let remaining = bytes;
-    const capacity =
-      this.capacity > 0 ? this.capacity : this.rateBytesPerSecond;
-
-    while (remaining > 0) {
-      if (signal?.aborted) {
-        throw new DOMException('Aborted', 'AbortError');
-      }
-
-      this.refill();
-      const target = Math.min(remaining, capacity);
-      if (this.tokens < target) {
-        const deficit = target - this.tokens;
-        const waitMs = Math.ceil((deficit / this.rateBytesPerSecond) * 1000);
-        await sleep(waitMs, signal);
-        continue;
-      }
-
-      this.tokens -= target;
-      remaining -= target;
-    }
-  }
-}
-
-function sleep(ms: number, signal?: AbortSignal): Promise<void> {
-  return new Promise((resolve, reject) => {
-    if (signal?.aborted) {
-      reject(new DOMException('Aborted', 'AbortError'));
-      return;
-    }
-
-    const timer = setTimeout(() => {
-      if (signal) {
-        signal.removeEventListener('abort', onAbort);
-      }
-      resolve();
-    }, ms);
-
-    const onAbort = () => {
-      clearTimeout(timer);
-      if (signal) {
-        signal.removeEventListener('abort', onAbort);
-      }
-      reject(new DOMException('Aborted', 'AbortError'));
-    };
-
-    if (signal) {
-      signal.addEventListener('abort', onAbort, { once: true });
-    }
-  });
-}
-
-function createThrottledStream(
-  blob: Blob,
-  limiter: UploadRateLimiter,
-  signal?: AbortSignal
-): ReadableStream<Uint8Array> | Blob {
-  if (!limiter.isEnabled() || typeof blob.stream !== 'function') {
-    return blob;
-  }
-
-  const reader = blob.stream().getReader();
-
-  return new ReadableStream<Uint8Array>({
-    async pull(controller) {
-      if (signal?.aborted) {
-        controller.error(new DOMException('Aborted', 'AbortError'));
-        return;
-      }
-
-      const { value, done } = await reader.read();
-      if (done) {
-        controller.close();
-        return;
-      }
-
-      if (value) {
-        await limiter.consume(value.byteLength, signal);
-        controller.enqueue(value);
-      }
-    },
-    cancel() {
-      void reader.cancel();
-    },
-  });
-}
-
+// Worker Context
 const workerContext = self as unknown as {
   postMessage: (event: UploadWorkerEvent) => void;
   onmessage: ((event: MessageEvent<UploadWorkerCommand>) => void) | null;
 };
 
-const queue = new UploadQueue<UploadTask>();
+// 상수
 const PROCESSING_HEARTBEAT_MS = 500;
+
+// 상태 변수
+const queue = new UploadQueue<UploadTask>();
+const rateLimiter = new UploadRateLimiter();
 let networkStatus: StreamHealthState = 'HEALTHY';
 let uploadRateBps: number | null = null;
 let streamingActive = true;
 let processing = false;
 let timerId: ReturnType<typeof setTimeout> | null = null;
 let activeAbortController: AbortController | null = null;
-const rateLimiter = new UploadRateLimiter();
 
+// 이벤트 헬퍼
 function postEvent(event: UploadWorkerEvent): void {
   workerContext.postMessage(event);
 }
@@ -179,6 +48,7 @@ function clearTimer(): void {
   }
 }
 
+// 업로드 정책
 function isUploadAllowed(): boolean {
   if (!streamingActive) {
     return true;
@@ -195,6 +65,7 @@ function isUploadAllowed(): boolean {
   return true;
 }
 
+// 스케줄링
 function scheduleNext(): void {
   clearTimer();
 
@@ -221,94 +92,7 @@ function scheduleNext(): void {
   }, delay);
 }
 
-async function uploadToS3(
-  s3Url: string,
-  contentType: string,
-  blob: Blob,
-  signal: AbortSignal
-): Promise<void> {
-  const body = createThrottledStream(blob, rateLimiter, signal);
-  const response = await fetch(s3Url, {
-    method: 'PUT',
-    headers: {
-      'Content-Type': contentType,
-    },
-    body,
-    signal,
-    // @ts-expect-error: duplex is not yet in the RequestInit type definition but required for streaming bodies
-    duplex: 'half',
-  });
-
-  if (!response.ok) {
-    throw new Error(`S3 업로드 실패: ${response.status}`);
-  }
-}
-
-async function performUpload(task: UploadTask): Promise<{
-  remoteSegmentId: string;
-  s3Url: string;
-}> {
-  if (!task.remoteSegmentId || !task.s3Url) {
-    // 2. Presigned URL 요청
-
-    const { segmentId: backendSegmentId, s3Url } = await postPresignedUrl(
-      task.sessionId,
-      {
-        sequence: task.sequence,
-        video_start_ms: task.segment.start_media_time,
-        video_end_ms: task.segment.end_media_time,
-        content_type: task.contentType,
-      }
-    );
-
-    task.remoteSegmentId = backendSegmentId;
-    task.s3Url = s3Url;
-  }
-
-  if (!task.s3Uploaded && task.s3Url) {
-    activeAbortController = new AbortController();
-    try {
-      await uploadToS3(
-        task.s3Url,
-        task.contentType,
-        task.blob,
-        activeAbortController.signal
-      );
-
-      task.s3Uploaded = true;
-    } finally {
-      activeAbortController = null;
-    }
-  }
-
-  if (!task.completeNotified && task.remoteSegmentId) {
-    await postUploadComplete(task.sessionId, task.remoteSegmentId);
-
-    task.completeNotified = true;
-  }
-
-  if (!task.logsUploaded && task.remoteSegmentId && task.s3Url) {
-    if (task.logs.length > 0) {
-      await postInputLogs(
-        task.sessionId,
-        task.remoteSegmentId,
-        task.s3Url,
-        task.logs
-      );
-    }
-    task.logsUploaded = true;
-    task.logs = [];
-  }
-
-  // 6. 로컬 저장소 정리 (성공 시) (processQueue에서 처리하므로 메시지만 전송해도 되지만, 리턴값이 필요함)
-  postMessage({ type: 'UPLOAD_SUCCESS', segmentId: task.segment.segment_id });
-
-  return {
-    remoteSegmentId: task.remoteSegmentId ?? task.segment.segment_id,
-    s3Url: task.s3Url ?? '',
-  };
-}
-
+// 큐 처리
 async function processQueue(): Promise<void> {
   if (processing || !isUploadAllowed()) {
     return;
@@ -324,13 +108,12 @@ async function processQueue(): Promise<void> {
   processing = true;
   queue.markInFlight(nextItem.key);
   const startedAt = new Date().toISOString();
+
   postEvent({
     type: 'segment-processing',
-    payload: {
-      localSegmentId: nextItem.key,
-      startedAt,
-    },
+    payload: { localSegmentId: nextItem.key, startedAt },
   });
+
   const heartbeatId = setInterval(() => {
     postEvent({
       type: 'segment-processing',
@@ -342,7 +125,18 @@ async function processQueue(): Promise<void> {
   }, PROCESSING_HEARTBEAT_MS);
 
   try {
-    const result = await performUpload(nextItem.payload);
+    const result = await performUpload({
+      task: nextItem.payload,
+      rateLimiter,
+      getAbortController: () => {
+        activeAbortController = new AbortController();
+        return activeAbortController;
+      },
+      clearAbortController: () => {
+        activeAbortController = null;
+      },
+    });
+
     queue.markDone(nextItem.key);
     postEvent({
       type: 'segment-uploaded',
@@ -360,10 +154,7 @@ async function processQueue(): Promise<void> {
       queue.markDone(nextItem.key);
       postEvent({
         type: 'segment-failed',
-        payload: {
-          localSegmentId: nextItem.key,
-          reason,
-        },
+        payload: { localSegmentId: nextItem.key, reason },
       });
     }
   } finally {
@@ -374,6 +165,7 @@ async function processQueue(): Promise<void> {
   }
 }
 
+// 커맨드 핸들러
 function enqueueSegment(payload: UploadWorkerSegmentPayload): void {
   const localSegmentId = payload.segment.segment_id;
 
@@ -417,6 +209,7 @@ function resetQueue(): void {
   notifyQueueSize();
 }
 
+// 메시지 핸들러
 workerContext.onmessage = (event: MessageEvent<UploadWorkerCommand>) => {
   const message = event.data;
